+++ variable-assembly-01/vm.h
@@ -5,7 +5,7 @@ typedef void *Label;
 typedef union Cell
 {
   Label inst;
-  long i;
+  long long i;
   union Cell *target;
 } Cell, Inst;
 
@@ -38,8 +38,8 @@ struct block_count *block_insert(Inst *ip);
 void vm_print_profile(FILE *file);
 
 /* vm type-specific support functions */
-void genarg_i(Inst **vmcodepp, long i);
-void printarg_i(long i);
+void genarg_i(Inst **vmcodepp, long long i);
+void printarg_i(long long i);
 void genarg_target(Inst **vmcodepp, Inst *target);
 void printarg_target(Inst *target);
 void printarg_Cell(Cell i);
